[X] send tree changed events on parenting
so thye will be triggered in all cases
