+++ Xamarin.Forms.Core/Xamarin.Forms.Core.csproj
@@ -20,6 +20,9 @@
 		<PackageReference Include="System.Dynamic.Runtime" Version="4.3.0" />
 		<PackageReference Include="System.ValueTuple" Version="4.5.0" />
 	</ItemGroup>	
+	<ItemGroup>
+	  <Folder Include="Xaml\Diagnostics\" />
+	</ItemGroup>
 	<PropertyGroup>
     <GitInfoReportImportance>high</GitInfoReportImportance>
   </PropertyGroup>
